Added kernel file name generation from expr type.
For each AssignmentExpression type, a unique kernel filename is
now generated if the -DREUSE_KERNELS option is set, and previously
generated kernels are reused if they exist.

Note that no precautions are taken to ensure that the kernels are
valid in any way. If the file exists, it will be loaded, and if
loading fails, the program will probably abort. A later release
should implement automatic recompilation if loading fails, and
possibly also generate different filenames if, for example, the
existing file cannot be read (due to access restrictions or the
like). Also, the test for file existence uses sys/stat.h, which
may not be portable.