Added recompilation of existing kernels if needed.
If -DREUSE_KERNELS is set and loading an existing kernel fails,
it will automatically be regenerated. Some errors in the ptx file
may leave the CUDA runtime API in a broken state, so that the next
call to cuModuleLoad may segfault. This should however not be a
problem in practice.